Test programs have been updated.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1490 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUTools++/Testprog/GuidedFilter/main.cc
@@ -30,7 +30,7 @@ main(int argc, char* argv[])
     cin >> a;
 
     Array<u_char>	b(a.size());
-    for (u_int i = 0; i < b.size(); ++i)
+    for (size_t i = 0; i < b.size(); ++i)
 	b[i] = a[i]; //+ 1;
 
     GuidedFilter<float>	gf(w, e);
@@ -45,8 +45,8 @@ main(int argc, char* argv[])
     cin >> A;
 
     Array2<Array<u_char> >	B(A.nrow(), A.ncol());
-    for (u_int i = 0; i < B.nrow(); ++i)
-	for (u_int j = 0; j < B.ncol(); ++j)
+    for (size_t i = 0; i < B.nrow(); ++i)
+	for (size_t j = 0; j < B.ncol(); ++j)
 	    B[i][j] = A[i][j];// + 1;
 
     GuidedFilter2<float>	gf2(w, w, e);
